LayoutTests/fast/viewport/viewport-43-expected.txt

CONSOLE MESSAGE: line 3: Viewport width or height set to physical device width, try using "device-width" constant instead for future compatibility.
viewport size 64x70 scale 5.000000 with limits [5.000000, 5.000000]
